Rename Gitlab::FileEditor to Gitlab::Satellite::EditFileAction
